do not allow editing templates that are in use
