+++ tutor/src/models/grading/templates.js
@@ -29,6 +29,7 @@ class GradingTemplate extends BaseModel {
   @field default_due_time = '21:00';
   @field default_due_date_offset_days = 7;
   @field default_close_date_offset_days = 7;
+  @field has_task_plans;
   @observable map;
 
   @observable map = {};
